Moved wasCheckCached test from Sensor to Status
