+++ tests/TestCase/Heartbeat/SensorTest.php
@@ -73,37 +73,4 @@ class SensorTest extends TestCase
         $this->assertEquals('2017-03-30 12:45:37', $status->getLastExecuted());
         $this->assertEquals(1, $status->getSeverity());
     }
-
-    /**
-     * Tests whether the check result was fetched from cache or by running the check now
-     *
-     * @return void
-     * @covers Sensor\Status::wasCheckCached
-     */
-    public function testWasCheckCached()
-    {
-        Chronos::setTestNow('2017-03-30 12:45:37');
-        $sensorName = 'Dummy Sensor';
-        $sensorConfig = [
-            'enabled' => true,
-            'severity' => 1,
-            'class' => DummySensor::class,
-            'cached' => "+1 seconds"
-        ];
-        $sensorConfig = new Sensor\Config($sensorName, $sensorConfig);
-        $sensorClassName = $sensorConfig->getClass();
-
-        /** @var Sensor $sensor */
-        $sensor = new $sensorClassName($sensorConfig);
-        $status = $sensor->getStatus();
-
-        // Assert that result was not cached
-        $this->assertFalse($status->wasCheckCached());
-
-        // Get status again and assert that result was cached
-        /** @var Sensor $sensor */
-        $sensor = new $sensorClassName($sensorConfig);
-        $status = $sensor->getStatus();
-        $this->assertTrue($status->wasCheckCached());
-    }
 }
